+++ src/subscription/application/services/subscription-manager.service.ts
@@ -2,10 +2,10 @@ import { Subscription } from "src/subscription/domain/entities/subscription.enti
 import { CreateSubscriptionDto } from "src/weather/application/dto/create-subscription.dto";
 import { ConflictException, Inject, Injectable } from "@nestjs/common";
 import { v4 as uuidv4 } from "uuid";
-import {
-  ISubscriptionRepository,
-  SUBSCRIPTION_REPOSITORY,
-} from "@/subscription/infrastructure/repositories/subscription.repository.interface";
+import { SUBSCRIPTION_QUERY_REPOSITORY } from "@/subscription/infrastructure/repositories/subscription-query.repository.interface";
+import { SUBSCRIPTION_COMMAND_REPOSITORY } from "@/subscription/infrastructure/repositories/subscription-command.repository.interface";
+import { SubscriptionQueryRepository } from "@/subscription/infrastructure/repositories/subscription-query.repository";
+import { SubscriptionCommandRepository } from "@/subscription/infrastructure/repositories/subscription-command.repository";
 
 export class SubscriptionFactory {
   create(dto: CreateSubscriptionDto, token: string): Subscription {
@@ -16,28 +16,29 @@ export class SubscriptionFactory {
 @Injectable()
 export class SubscriptionManager {
   constructor(
-    @Inject(SUBSCRIPTION_REPOSITORY)
-    private readonly repo: ISubscriptionRepository,
+    @Inject(SUBSCRIPTION_QUERY_REPOSITORY)
+    private readonly queryRepo: SubscriptionQueryRepository,
+    @Inject(SUBSCRIPTION_COMMAND_REPOSITORY)
+    private readonly commandRepo: SubscriptionCommandRepository,
     private readonly factory: SubscriptionFactory,
   ) {}
 
   async subscribe(dto: CreateSubscriptionDto, token: string): Promise<Subscription> {
-    const isSubbed = await this.repo.isEmailSubscribed(dto.email, dto.city);
+    const isSubbed = await this.queryRepo.isEmailSubscribed(dto.email, dto.city);
     if (isSubbed) {
       throw new ConflictException("Email already subscribed");
     }
 
     const subscription = this.factory.create(dto, token);
-    await this.repo.create(subscription);
-
+    await this.commandRepo.create(subscription);
     return subscription;
   }
 
   async confirm(token: string): Promise<void> {
-    await this.repo.confirmSubscription(token);
+    await this.commandRepo.confirmSubscription(token);
   }
 
   async unsubscribe(token: string): Promise<void> {
-    await this.repo.unsubscribe(token);
+    await this.commandRepo.unsubscribe(token);
   }
 }
